lib: lib-signals - Make sure handled signals are not blocked.
Dovecot processes not started by the Dovecot master inherit the sigprocmask()
signal mask from the respective external parent process. This is for example
true for Dovecot LDA. This means that some signals could be blocked unexpectedly
and the lib-signals API does not account for this eventuality. This is fixed by
making sure signals for which a handler is installed are not blocked by the
signal mask.